requirements.txt
setuptools
twine
build
black
python-dotenv